Introduce buffer interface to allow multiple types of buffer impls
